added formspree redirection in contact form
Before, clicking on the submit button of the contact form would redirect to the mail client of the user. Then the user had to send the contact form information through his email account. But now, clicking on the submit button will directly send the contact form information to mushfiqurrahman6069@gmail.com using Formspree.